% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three_item_scale.R
\name{alpha_3item_combo}
\alias{alpha_3item_combo}
\title{Find the 3-item short scale (criterion = Cronbach's alpha)}
\usage{
alpha_3item_combo(x)
}
\arguments{
\item{x}{The vector with all the items in the questionnaire. All columns must be numeric.}
}
\description{
This function tests all 3-item combinations possible in your questionnaire and returns the combo that has the highest internal consistency (based on Cronbach's alpha)
}
\examples{
library(psych) #load the psych package
data("bfi") #read in the bfi dataset from the psych package
test <- as.data.frame(cbind(bfi$A1,bfi$A2,bfi$A3,bfi$A4,bfi$A5)) #save the 5 agreeableness items as a dataframe
alpha_3item_combo(test) #run the function on the dataframe
}
\keyword{3-item}
\keyword{alpha}
\keyword{brief}
\keyword{cronbach}
\keyword{questionnaire}
\keyword{scale}
