% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfI_alpha.R
\name{bfi_alpha}
\alias{bfi_alpha}
\title{A Function to calculate bfi trait score reliability}
\usage{
bfi_alpha(data1)
}
\arguments{
\item{data}{The dataframe. Note that the bfi items must be labelled correctly}
}
\description{
This function takes a dataframe with Big Five Inventory 2 items labelled bfi_1, bfi_2, etc. and returns alpha reliability for each trait.
}
\examples{
bfi_alpha(data)
bfi_alpha()
}
\keyword{bfi,}
\keyword{reliability}
\keyword{scores,}
