% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colchartonum.R
\name{colchartonum}
\alias{colchartonum}
\title{A Colchartonum Function}
\usage{
colchartonum(data)
}
\arguments{
\item{data}{The dataframe.}
}
\description{
This function changes all columns that are characters to numeric.
}
\examples{
library(car)
#Load Prestige of Canadian Occupations Data from car package:
data <- Prestige 
data$income <- as.character(data$income)
head(data)
str(data)
data <- colchartonum(data) 
#Re-run str(data) to make sure all characters were converted to numeric--should have changed one variable
str(data)
colchartonum()
}
\keyword{character,}
\keyword{class,}
\keyword{data}
\keyword{numeric}
\keyword{structure,}
