% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colfacttonum.R
\name{colfacttonum}
\alias{colfacttonum}
\title{A Colfacttonum Function}
\usage{
colfacttonum(data)
}
\arguments{
\item{data}{The dataframe.}
}
\description{
This function changes all columns that are factors to numeric.
}
\examples{
library(car)
#Load Prestige of Canadian Occupations Data from car package:
data <- Prestige 
data$income <- as.character(data$income)
head(data)
str(data)
data <- colfacttonum(data) 
#Re-run str(data) to make sure all factors were converted to numeric--should have changed one variable
str(data)
colfacttonum()
}
\keyword{class,}
\keyword{data}
\keyword{factors,}
\keyword{numeric}
\keyword{structure,}
