% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colinttonum.R
\name{colinttonum}
\alias{colinttonum}
\title{A Colinttonum Function}
\usage{
colinttonum(data)
}
\arguments{
\item{data}{The dataframe.}
}
\description{
This function changes all columns that are integers to numeric.
}
\examples{
library(car)
#Load Prestige of Canadian Occupations Data from car package:
data <- Prestige 
data$income <- as.character(data$income) 
head(data)
str(data)
data <- colinttonum(data)
#Re-run str(data) to make sure all integers were converted to numeric--should have changed one variable
str(data)
colinttonum()
}
\keyword{class,}
\keyword{data}
\keyword{integer,}
\keyword{numeric}
\keyword{structure,}
