% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_cv.R
\name{corr_cv}
\alias{corr_cv}
\title{A Cross-Validated Correlation Function}
\usage{
corr_cv(x, y)
}
\arguments{
\item{x}{A single vector.}

\item{y}{A single vector.}
}
\description{
This function allows you to calculate cross-validated correlations.
}
\examples{
corr_cv(got$`Book Intro Chapter`, got$`Death Chapter`)
}
\keyword{correlation}
\keyword{cross-validation}
