% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csus_mean_full.R
\name{csus_mean_full}
\alias{csus_mean_full}
\title{csus_mean_full Function}
\usage{
csus_mean_full(data)
}
\arguments{
\item{data}{The dataframe.}
}
\description{
This function calculates the scale mean for the full 42 items of the CSUS
}
\examples{
csus_mean_full(data)
csus_mean_full()
}
\keyword{mean}
