% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csus_mean_short.R
\name{csus_mean_short}
\alias{csus_mean_short}
\title{csus_mean_short Function}
\usage{
csus_mean_short(data)
}
\arguments{
\item{data}{The dataframe.}
}
\description{
This function calculates the scale mean for the short 18 items of the CSUS. Short scale items that corresponds to the full scale includes: Item# 1, 3, 7, 12, 14, 18, 19, 21, 24, 25, 27, 30, 32, 33, 34, 36, 38, 40.
}
\examples{
csus_mean_short(data)
csus_mean_short()
}
\keyword{mean}
