% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csus_recode.R
\name{csus_recode}
\alias{csus_recode}
\title{csus_recode Function}
\usage{
csus_recode(data)
}
\arguments{
\item{data}{The dataframe.}
}
\description{
This function recodes items in the CSUS - Items #5 12, 25, 35, 39, 41 - that needs to be reverse coded. The csus_recode function will recode into the same variable as opposed to creating a new variable.
}
\examples{
csus_recode(data)
csus_recode()
}
\keyword{recode}
