% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{descriptives}
\alias{descriptives}
\title{A Descriptives Function}
\usage{
descriptives(data, si.digits = 3)
}
\arguments{
\item{data}{The dataframe. Note that all variables must be numeric.}

\item{si.digits}{The number of digits to round to. Defaults to 3 decimal places.}
}
\description{
This function allows you to examine your favorite descriptive statistics.
}
\examples{
library(car)
#Load Prestige of Canadian Occupations Data from car package:
data <- Prestige
head(data)
str(data)
#Prepare data--all variables must be numeric
data <- colfacttonum(data)
data <- colinttonum(data)
descriptives(data)
descriptives()
}
\keyword{cases,}
\keyword{descriptive}
\keyword{deviation,}
\keyword{error,}
\keyword{kurtosis}
\keyword{mean,}
\keyword{missing}
\keyword{skew,}
\keyword{standard}
\keyword{stats,}
\keyword{valid}
\keyword{vases,}
