% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_file_import.R
\name{mat_file_import}
\alias{mat_file_import}
\title{Importing list of Matlab files}
\usage{
mat_file_import(pattern)
}
\arguments{
\item{pattern}{A pattern that is unique to a set of matlab files in your working directory in the form of a string}
}
\value{
list of files
}
\description{
This function allows you to import a list of specific .mat files in your working directory by specifying a particular pattern that is in the matlab file name.
}
\examples{
mat_file_import("_test1_")
}
\keyword{.mat,}
\keyword{import}
\keyword{list,}
\keyword{matlab,}
\keyword{pattern,}
