% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_blue.R
\name{plot_blue}
\alias{plot_blue}
\title{A Blue Plot Function}
\usage{
plot_blue(x, y)
}
\arguments{
\item{x}{A single vector.}

\item{y}{A single vector.}
}
\description{
This function allows you to create a plot in blue.
}
\examples{
plot_blue(got$`Book Intro Chapter`, got$`Death Chapter`)
}
\keyword{blue}
\keyword{plot}
