% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reputation_model.R
\name{rep_consensus}
\alias{rep_consensus}
\title{Reputation Consensus Model}
\usage{
rep_consensus(data, p1_reports, p2_reports, n_triads = length(p1_reports),
  n_p1s_per_p2s = 1, n_p2s_per_p1s = 1)
}
\arguments{
\item{data}{The dataframe that contains P1 and P2 ratings.
Data should wide, with a row for every group of participants.
At a minimum, it must contain two columns: one for P1 reports and one for P2 reports.}

\item{p1_reports}{The column(s) that contain P1 reports,
or ratings made by the person that knows the target directly.
If more than one is supplied, the order must match the other
rating types.}

\item{p2_reports}{The column(s) that contain P1 reports,
or ratings made by the person that knows the target indirectly through the corresponding P1.
If more than one is supplied, the order must match the other
rating types.}

\item{n_triads}{The number of exchangeable triads in each group. By default, this is determined by
counting the number of P1 reports. This parameter rarely needs to be changed.}

\item{n_p1s_per_p2s}{The number of P1s for every P2. This defaults to 1.
Currently, only values of 1 are supported.}

\item{n_p2s_per_p1s}{The number of P2s for every P1;. This defaults to 1.
Currently, only values of 1 are supported.}
}
\description{
This takes a vector of P1 and P2 reports and fits
a model estimating the possible hearsay reputation parameters.
Those parameters are:
hc - hearsay consensus; the correlation between P1(T) & P2(T)
int_p1 - Intercept for P1(T)
int_p2 - Intercept for P2(T)
v_p1 - variance for P1(T)
v_p2 - variance for P2(T)
p1_p2_rel_el - P1-P2 Relative Elevation (i.e., Mean P1(T) - Mean P2(T))
}
\details{
If n exchangeable triads > 1:
rec - direct reciprocity; the correlation between opposit P1(T)s (e.g., A(C) <-> C(A))
h - hearsay reciprocity; the correlation between exchangeable P2(T)s (e.g., B(C) <-> D(A))
m - unnamed parameter; The correlation between P2(T) and the opposite P1(T) in a group. (e.g., B(C) <-> C(A))

The function can handle up to n exchangeable triads. It returns a fitted lavaan
object. See lavaan documentation for more.
}
\examples{
data("rep_sim_data")
          rep_consensus(data = rep_sim_data,
                        p1_reports = c("A_C_agreeableness", "C_A_agreeableness"),
                        p2_reports = c("B_C_agreeableness", "D_A_agreeableness"))
}
