% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_bfi.R
\name{score_bfi}
\alias{score_bfi}
\title{A Function to calculate bfi trait scores}
\usage{
score_bfi(data1)
}
\arguments{
\item{data}{The dataframe. Note that the bfi items must be labelled correctly}
}
\description{
This function takes a dataframe with Big Five Inventory 2 items labelled bfi_1, bfi_2, etc. and reverses necessary items and scores the five trait domains.
}
\examples{
bfi_alpha(data)
score_bfi()
}
\keyword{bfi,}
\keyword{reliability}
\keyword{scores,}
