% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE.R
\name{SE}
\alias{SE}
\title{A Standard Error Function}
\usage{
SE(x)
}
\arguments{
\item{x}{A single vector.}
}
\description{
This function allows you to calculate standard error.
}
\examples{
SE(got$`Death Chapter`)
}
\keyword{error}
\keyword{standard}
