% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_wide_format.R
\name{se_wide_format}
\alias{se_wide_format}
\title{standard Error of Wide Format Data}
\usage{
se_wide_format(data, na.rm = FALSE)
}
\arguments{
\item{data}{The dataframe. Note that the dataframe should be in wide format where the first column represents individual subjects with only one row per subject and each additional column represents a specific test or task.}

\item{na.rm}{Default set to FALSE. If TRUE, missing data will be removed from the data frame before calculations proceed.}
}
\value{
dataframe
}
\description{
This function allows you to iteratively obtain standard error for each individual column in a wide format dataframe in which each row represents a particular subject and each column represents a specific test or task. This format generally displays summary data for particular tests or tasks where each individual cell is a mean or median of several trials of that particular test or task.
}
\examples{
subject_num <- c(1, 2, 3, 4) task1 <- c(99, 87, 82, 76) task2 <- c(67, 74, 68, 88) total <- as.data.frame(cbind(subject_num, task1, task2)) se_wide_format(total)
se_wide_format()
}
\keyword{data}
\keyword{error,}
\keyword{format,}
\keyword{standard}
\keyword{summary}
\keyword{wide}
